requirements.txt
asgiref==3.5.1
boto3==1.22.9
botocore==1.25.9
Django==4.0.3
djangorestframework==3.13.1
jmespath==1.0.0
mysqlclient==2.1.0
boto3==1.22.7
botocore==1.25.7
Django==4.0.4
django-storages==1.12.3

jmespath==1.0.0
Pillow==9.1.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
wheel==0.37.1
django-cognito-jwt==0.0.3
